New permission approach

+++ src/main/java/com/issuetracker/pages/project/ProjectDetail.java
@@ -3,19 +3,19 @@ package com.issuetracker.pages.project;
 import com.issuetracker.pages.layout.PageLayout;
 import com.issuetracker.model.Component;
 import com.issuetracker.model.CustomField;
-import com.issuetracker.model.Permission;
-import com.issuetracker.model.PermissionType;
 import com.issuetracker.model.Project;
 import com.issuetracker.model.ProjectVersion;
+import com.issuetracker.model.TypeId;
 import com.issuetracker.pages.component.component.ComponentListView;
 import com.issuetracker.pages.component.customField.CustomFieldListView;
-import com.issuetracker.pages.component.permission.PermissionsForm;
 import com.issuetracker.pages.component.version.VersionListView;
-import com.issuetracker.pages.component.permission.PermissionsListView;
+import com.issuetracker.pages.permissions.AccessDenied;
+import com.issuetracker.pages.permissions.ProjectPermission;
+import com.issuetracker.pages.validator.ProjectNameValidator;
 import com.issuetracker.service.api.ProjectService;
+import com.issuetracker.service.api.SecurityService;
 import static com.issuetracker.web.Constants.*;
-import static com.issuetracker.web.security.KeycloakAuthSession.*;
-import static com.issuetracker.web.security.PermissionsUtil.*;
+import java.util.ArrayList;
 import org.apache.wicket.markup.html.basic.Label;
 import org.apache.wicket.markup.html.form.Form;
 import org.apache.wicket.model.CompoundPropertyModel;
@@ -28,10 +28,11 @@ import java.util.List;
 import org.apache.wicket.markup.html.WebMarkupContainer;
 import org.apache.wicket.markup.html.form.TextField;
 import org.apache.wicket.model.Model;
-import java.util.ArrayList;
+import org.apache.wicket.markup.html.form.RequiredTextField;
 import org.apache.wicket.markup.html.link.Link;
 import org.apache.wicket.request.flow.RedirectToUrlException;
 import org.apache.wicket.util.string.StringValue;
+import org.jboss.logging.Logger;
 
 /**
  *
@@ -39,67 +40,98 @@ import org.apache.wicket.util.string.StringValue;
  */
 public class ProjectDetail extends PageLayout {
 
+    private final Logger log = Logger.getLogger(ProjectDetail.class);
+    
     @Inject
     private ProjectService projectService;
+    @Inject
+    private SecurityService securityService;
     
+    private final Form<String> renameForm;
     private final Form<ProjectVersion> projectVersionForm;
     private final Form<Component> projectComponentForm;
     private final Form<CustomField> projectCustomFieldForm;
     private Link deleteProject;
+    private Link projectPermissions;
     private List<ProjectVersion> projectVersionList;    
     private List<Component> projectComponentList;    
     private List<CustomField> projectCustomFieldsList;    
-    private List<Permission> projectPermissionsList;    
     private TextField versionTextField;
     private TextField componentTextField;
     private TextField customFieldTextField;
-    private Project project = new Project();
+    private Project project;
     private String stringVersion, stringComponent, stringCustomField;
-    private WebMarkupContainer wmcPermission;
-    private PermissionsForm<Permission> permissionForm;
+    
+    List<String> permittedActions = new ArrayList<>();
     
     @Override
     public void onConfigure() {
         super.onConfigure();
-        boolean hasEditPermissions = hasPermissionsProject(project, PermissionType.edit);
         
-        projectVersionForm.setVisible(hasEditPermissions);
-        projectComponentForm.setVisible(hasEditPermissions);
-        projectCustomFieldForm.setVisible(hasEditPermissions);
-        wmcPermission.setVisible(hasEditPermissions);
-        permissionForm.setVisible(hasEditPermissions);
-        
-        boolean hasDeletePermission = 
-                isProjectOwner(project) ||
-                isUserInAppRole("project.delete") ||
-                isSuperUser();
-        deleteProject.setVisible(hasDeletePermission);
+        renameForm.setVisible(permittedActions.contains(roles.getProperty("it.project.rename")));
+        projectVersionForm.setVisible(permittedActions.contains(roles.getProperty("it.project.versions")));
+        projectComponentForm.setVisible(permittedActions.contains(roles.getProperty("it.project.components")));
+        projectCustomFieldForm.setVisible(permittedActions.contains(roles.getProperty("it.project.custom.fields")));
+        deleteProject.setVisible(permittedActions.contains(roles.getProperty("it.project.delete")));
+        projectPermissions.setVisible(permittedActions.contains(roles.getProperty("it.project.permissions")));
     }
     
     public ProjectDetail(PageParameters parameters) {
-        StringValue projectId = parameters.get("project");
-        if (projectId.equals(StringValue.valueOf((String)null))) {
+        StringValue stringProjectId = parameters.get("project");
+        if (stringProjectId.equals(StringValue.valueOf((String)null))) {
+            log.warn("Page parameters doesn't contain project id. Redirecting to Home page.");
             throw new RedirectToUrlException(HOME_PAGE);
         }
+        Long projectId = stringProjectId.toLong();
+        project = projectService.getProjectById(projectId);
         
-        project = projectService.getProjectById(projectId.toLong());
+        if (project == null) {
+            log.warn("Project with given id doesn't exist. Redirecting to Home page.");
+            throw new RedirectToUrlException(HOME_PAGE);
+        }
+        if (!securityService.canUserPerformAction(TypeId.project, projectId, roles.getProperty("it.project.browse"))) {
+            setResponsePage(AccessDenied.class);
+        }   
+        permittedActions = securityService.getPermittedActionsForUserAndItem(TypeId.project, projectId);
         
         projectVersionList = project.getVersions();
         projectComponentList = project.getComponents();
         projectCustomFieldsList = project.getCustomFields();
-        projectPermissionsList = new ArrayList<>(project.getPermissions());
         
-        add(new Label("name", project.getName()));
+        add(new Label("name", new PropertyModel<String>(this, "project.name")));
+        
+        renameForm = new Form<String>("renameForm") {
+            @Override
+            protected void onSubmit() {
+                checkPermission("it.project.rename");
+                projectService.update(project);
+            }
+        };
+        renameForm.add(new RequiredTextField<>("rename", new PropertyModel<String>(this, "project.name")).add(new ProjectNameValidator()));
+        add(renameForm);
+        
         add(new Label("summary", project.getSummary()));
         deleteProject = new Link("delete") {
 
             @Override
             public void onClick() {
+                checkPermission("it.project.delete");
                 projectService.remove(project);
                 setResponsePage(ListProjects.class);
             }
         };
         add(deleteProject);
+        
+        projectPermissions = new Link("permissions") {
+            @Override
+            public void onClick() {
+                checkPermission("it.project.permissions");
+                PageParameters pageParameters = new PageParameters();
+                pageParameters.add("project", project.getId());
+                setResponsePage(ProjectPermission.class, pageParameters);
+            }
+        };
+        add(projectPermissions);
         
 //<editor-fold defaultstate="collapsed" desc="versions">
         IModel<List<ProjectVersion>> versionModel = new CompoundPropertyModel<List<ProjectVersion>>(projectVersionList) {
@@ -119,6 +151,7 @@ public class ProjectDetail extends PageLayout {
         projectVersionForm = new Form<ProjectVersion>("projectVersionForm") {
             @Override
             protected void onSubmit() {
+                checkPermission("it.project.versions");
                 ProjectVersion projectVersion = new ProjectVersion();
                 projectVersion.setName(versionTextField.getInput());
                 stringVersion = null; //this will clear version textfield
@@ -149,6 +182,7 @@ public class ProjectDetail extends PageLayout {
         projectComponentForm = new Form<Component>("projectComponentForm") {
             @Override
             protected void onSubmit() {
+                checkPermission("it.project.components");
                 Component component = new Component();
                 component.setName(componentTextField.getInput());
                 stringComponent = null; //this will clear component textfield
@@ -179,6 +213,7 @@ public class ProjectDetail extends PageLayout {
         projectCustomFieldForm = new Form<CustomField>("projectCustomFieldForm") {
             @Override
             protected void onSubmit() {
+                checkPermission("it.project.custom.fields");
                 CustomField customField = new CustomField();
                 customField.setCfName(customFieldTextField.getInput());
                 stringCustomField = null; //this will clear custom field text field
@@ -190,24 +225,6 @@ public class ProjectDetail extends PageLayout {
         projectCustomFieldForm.add(customFieldTextField);
         add(projectCustomFieldForm);
 //</editor-fold>
-        
-//<editor-fold defaultstate="collapsed" desc="permissions">
-        IModel<List<Permission>> permissionsModel = new CompoundPropertyModel<List<Permission>>(projectPermissionsList) {
-            @Override
-            public List<Permission> getObject() {
-                return projectPermissionsList;            
-            }
-        };
-        PermissionsListView<Permission> permissionsListView = new PermissionsListView<>("permissionsListView", permissionsModel, new Model<>(project));
-        wmcPermission = new WebMarkupContainer("wmcPermission");
-        wmcPermission.add(permissionsListView);
-        wmcPermission.setOutputMarkupId(true);
-        add(wmcPermission);
-        
-        permissionForm = new PermissionsForm("permissionForm", permissionsModel, new Model<>(project));
-        add(permissionForm);
-//</editor-fold>
-        
     }
 
     public String getStringVersion() {
@@ -232,5 +249,19 @@ public class ProjectDetail extends PageLayout {
 
     public void setStringCustomField(String stringCustomField) {
         this.stringCustomField = stringCustomField;
+    }
+
+    public Project getProject() {
+        return project;
+    }
+
+    public void setProject(Project project) {
+        this.project = project;
+    }
+    
+    private void checkPermission(String name) {
+        if (!permittedActions.contains(roles.getProperty(name))) {
+            setResponsePage(AccessDenied.class);
+        }
     }
 }
